refactor: split out clis to separate packages
