+++ Cargo.toml
@@ -1,36 +1,7 @@
-[package]
-name = "dapptools"
-version = "0.1.0"
-edition = "2018"
-
 [workspace]
 members = [
     "utils",
     "seth",
     "dapp",
+    "dapptools",
 ]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dapp-utils = { path = "./utils" }
-ethers = "0.5"
-
-structopt = "0.3.22"
-eyre = "0.6.5"
-tokio = { version = "1.10.1", features = ["macros"] }
-rustc-hex = "2.1.0"
-serde_json = "1.0.67"
-rpassword = "5.0.1"
-evm = { version = "0.30.1" }
-ansi_term = "0.12.1"
-serde = "1.0.130"
-hex = "0.4.3"
-regex = { version = "1.5.4", default-features = false }
-svm = { package = "svm-rs", git = "https://github.com/roynalnaruto/svm-rs" }
-glob = "0.3.0"
-semver = "1.0.4"
-
-
-[patch.'crates-io']
-evm = { git = "https://github.com/gakonst/evm" }
