+++ dapptools/src/seth.rs
@@ -1,15 +1,82 @@
-use std::convert::TryFrom;
+mod opts;
+use opts::EthereumOpts;
+
+use seth::{Seth, SimpleSeth};
 
 use ethers::{
-    prelude::SignerMiddleware,
+    middleware::SignerMiddleware,
     providers::{Middleware, Provider},
     signers::Signer,
-    types::Address,
+    types::{Address, BlockId, BlockNumber, H256, U64},
 };
+use std::{convert::TryFrom, str::FromStr};
 use structopt::StructOpt;
 
-use dapptools::opts::{Opts, Subcommands};
-use dapptools::{Seth, SimpleSeth};
+#[derive(Debug, StructOpt)]
+#[structopt(about = "Perform Ethereum RPC calls from the comfort of your command line.")]
+pub enum Subcommands {
+    #[structopt(name = "--from-ascii")]
+    #[structopt(about = "convert text data into hexdata")]
+    FromAscii { text: String },
+    #[structopt(name = "--to-checksum-address")]
+    #[structopt(about = "convert an address to a checksummed format (EIP-55)")]
+    ToCheckSumAddress { address: Address },
+    #[structopt(name = "--to-bytes32")]
+    #[structopt(about = "left-pads a hex bytes string to 32 bytes)")]
+    ToBytes32 { bytes: String },
+    #[structopt(name = "block")]
+    #[structopt(
+        about = "Prints information about <block>. If <field> is given, print only the value of that field"
+    )]
+    Block {
+        #[structopt(help = "the block you want to query, can also be earliest/latest/pending", parse(try_from_str = parse_block_id))]
+        block: BlockId,
+        #[structopt(long, env = "SETH_FULL_BLOCK")]
+        full: bool,
+        field: Option<String>,
+        #[structopt(long = "--json", short = "-j")]
+        to_json: bool,
+        #[structopt(long, env = "ETH_RPC_URL")]
+        rpc_url: String,
+    },
+    #[structopt(name = "call")]
+    #[structopt(about = "Perform a local call to <to> without publishing a transaction.")]
+    Call {
+        #[structopt(help = "the address you want to query")]
+        address: Address,
+        sig: String,
+        args: Vec<String>,
+        #[structopt(long, env = "ETH_RPC_URL")]
+        rpc_url: String,
+    },
+    #[structopt(name = "send")]
+    #[structopt(about = "Publish a transaction signed by <from> to call <to> with <data>")]
+    SendTx {
+        #[structopt(help = "the address you want to transact with")]
+        to: Address,
+        #[structopt(help = "the function signature you want to call")]
+        sig: String,
+        #[structopt(help = "the list of arguments you want to call the function with")]
+        args: Vec<String>,
+        #[structopt(flatten)]
+        eth: EthereumOpts,
+    },
+}
+
+fn parse_block_id(s: &str) -> eyre::Result<BlockId> {
+    Ok(match s {
+        "earliest" => BlockId::Number(BlockNumber::Earliest),
+        "latest" => BlockId::Number(BlockNumber::Latest),
+        s if s.starts_with("0x") => BlockId::Hash(H256::from_str(s)?),
+        s => BlockId::Number(BlockNumber::Number(U64::from_str(s)?)),
+    })
+}
+
+#[derive(Debug, StructOpt)]
+pub struct Opts {
+    #[structopt(subcommand)]
+    pub sub: Subcommands,
+}
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
